requirements.txt
tensorflow==2.3.0
streamlit==1.18.0
numpy==1.20.3
pillow==9.4.0

# tensorflow==2.3.0
# streamlit==0.62.0
# numpy==1.16.5
# pillow==6.2.0

# tensorflow==2.11.0
# streamlit==1.17.0
# numpy==1.24.2
# pillow==9.4.0